# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1.0/events/create_event": {
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "Create Event",
                "operationId": "create_event_api_v1_0_events_create_event_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Event"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Event successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventKey"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1.0/events/get_event/{event_id}": {
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "Get Event",
                "operationId": "get_event_api_v1_0_events_get_event__event_id__put",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Event Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/EventWithId"
                                        },
                                        {
                                            "type": "null"
                                        }
                                    ],
                                    "title": "Response Get Event Api V1 0 Events Get Event  Event Id  Put"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found."
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1.0/events/add_attendees_to_event/{event_id}": {
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "Add Attendees To Event",
                "operationId": "add_attendees_to_event_api_v1_0_events_add_attendees_to_event__event_id__put",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Event Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                },
                                "title": "User Ids"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Event successfully updated."
                    },
                    "404": {
                        "description": "Event not found."
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1.0/events/delete_event/{event_id}": {
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "Delete Event",
                "operationId": "delete_event_api_v1_0_events_delete_event__event_id__put",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Event Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event successfully deleted."
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1.0/users/create_user": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Create User",
                "operationId": "create_user_api_v1_0_users_create_user_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserKey"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1.0/users/get_user/{user_id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Get User",
                "operationId": "get_user_api_v1_0_users_get_user__user_id__put",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/UserWithId"
                                        },
                                        {
                                            "type": "null"
                                        }
                                    ],
                                    "title": "Response Get User Api V1 0 Users Get User  User Id  Put"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1.0/users/delete_user/{user_id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Delete User",
                "operationId": "delete_user_api_v1_0_users_delete_user__user_id__put",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User successfully deleted."
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "408": {
                        "description": "Request timed out."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Event": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Time"
                    },
                    "location": {
                        "type": "string",
                        "title": "Location"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "attendees": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "integer"
                                },
                                "type": "array",
                                "uniqueItems": true
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Attendees"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "time",
                    "location",
                    "description"
                ],
                "title": "Event"
            },
            "EventKey": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "EventKey"
            },
            "EventWithId": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Time"
                    },
                    "location": {
                        "type": "string",
                        "title": "Location"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "attendees": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "integer"
                                },
                                "type": "array",
                                "uniqueItems": true
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Attendees"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "time",
                    "location",
                    "description",
                    "id"
                ],
                "title": "EventWithId"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "User": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "first_name",
                    "last_name",
                    "email"
                ],
                "title": "User"
            },
            "UserKey": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "UserKey"
            },
            "UserWithId": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "id"
                ],
                "title": "UserWithId"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}